fix the default va_arg expansion (in the realignment case) to not implicitly truncate the stack pointer to 32-bits on a 64-bit machine.
llvm-svn: 116169

+++ llvm/lib/CodeGen/SelectionDAG/LegalizeDAG.cpp
@@ -2699,7 +2699,7 @@ void SelectionDAGLegalize::ExpandNode(SDNode *Node,
                                            TLI.getPointerTy()));
 
       VAList = DAG.getNode(ISD::AND, dl, TLI.getPointerTy(), VAList,
-                           DAG.getConstant(-Align,
+                           DAG.getConstant(-(int64_t)Align,
                                            TLI.getPointerTy()));
     }
